Fix gradle builds on Mac OS X.
The checkstyles plugin configuration file being included could not be
found. This change uses the built-in configuration directory variable
to make it work cross-platform.

I had to do this because I was trying to fix the RSTAUI gradle build
to properly use Maven-based dependencies as discussed in
https://github.com/bobbylight/RSTAUI/issues/16 but I could not get
that to compile because of serialization problems that you fixed in
RSyntaxTextArea version 3.0.8-SNAPSHOT, but you had not pushed that
snapshot to the Sonatype OSSRH repository. By fixing this build, I am
now able to run `gradlew publishToMavenLocal` and it builds and
installs the latest snapshot into my local `.m2` Maven cache
repository, and I can proceed with fixing the RSTAUI build.